vector-search/requirements.txt
faiss-cpu==1.8.0
numpy==1.26.4
protobuf==5.27.3
